Refactor tests for twemproxy
