+++ twemproxy/tests/__init__.py
@@ -1,0 +1,3 @@
+# (C) Datadog, Inc. 2018
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
