+++ twemproxy/tests/common.py
@@ -1,0 +1,14 @@
+# (C) Datadog, Inc. 2018
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+
+import os
+
+from datadog_checks.utils.common import get_docker_hostname
+
+HERE = os.path.dirname(os.path.abspath(__file__))
+COMPOSE_FILE = os.path.join(HERE, 'compose', 'docker-compose.yaml')
+ROOT = os.path.dirname(os.path.dirname(HERE))
+TESTS_HELPER_DIR = os.path.join(ROOT, 'datadog_checks_tests_helper')
+HOST = get_docker_hostname()
+PORT = 6222
